swam-main/requirements.txt
PyQt5==5.15.9
pyqtdarktheme==1.1.0
numpy==1.25.0
scipy==1.11.1
matplotlib==3.7.2
pynmea2==1.18.0
pyserial==3.5
